Split up file to header/footer and sign up form component
