+++ src/components/Footer.vue
@@ -1,0 +1,99 @@
+<template>
+    <!-- FOOTER -->
+    <div class="ratingbar">
+    <div class="container">
+        <div class="row">
+            <div class="col-xs-12 col-sm-6 col-md-4 ratingbar__rating">
+                <a
+                    href="https://www.independer.nl/zorgverzekering/fi/zilveren-kruis-ac.aspx"
+                    class=""
+                    rel="noopener noreferrer"
+                    target="_blank"
+                >
+                    <img
+                        src="https://www.zilverenkruis.nl/-/media/zilverenkruis/images/shared/independer.png?rev=bb0602912ded4b16adc221025f3fc010&amp;h=80&amp;w=80&amp;la=nl-NL&amp;hash=39DBF59680B0FEECA97C6C786FC7DE67"
+                        class="ratingbar__icon"
+                        alt="Independer"
+                        title="Independer"
+                    />
+                </a>
+                <span>
+                    <span class="ratingbar__ratings-number"> 7,8 </span>
+                    <span class="ratingbar__ratings-text">
+                        op Independer
+                    </span>
+                </span>
+            </div>
+            <div class="col-xs-12 col-sm-6 col-md-4 ratingbar__rating">
+                <span>
+                    <span class="ratingbar__ratings-number">
+                        3,8 miljoen
+                    </span>
+                    <span class="ratingbar__ratings-text">
+                        klanten gingen u voor
+                    </span>
+                </span>
+            </div>
+            <div class="col-xs-12 col-sm-6 col-md-4 ratingbar__rating">
+                <a
+                    href="/consumenten/service/declareren/zilveren-kruis-app"
+                    class=""
+                >
+                    <img
+                        src="https://www.zilverenkruis.nl/-/media/zilverenkruis/images/app-logo/app_icon.png?rev=8665a8133399403b92b34eaada19bd46&amp;h=80&amp;w=80&amp;la=nl-NL&amp;hash=70AB973D0AB046A6EFA670CB38F2BF2E"
+                        class="ratingbar__icon"
+                        alt="Logo Zilveren Kruis-app"
+                    />
+                </a>
+                <span>
+                    <span class="ratingbar__ratings-number"> 4,5 </span>
+                    <span class="ratingbar__ratings-text">
+                        beoordeling in App Store
+                    </span>
+                </span>
+            </div>
+        </div>
+    </div>
+    </div>
+    <footer class="main-footer" role="contentinfo">
+        <div class="container main-footer__grid">
+            <ul
+                class="main-footer__links-section list list--nostyle list--horiz d-flex"
+            >
+                <li class="list__item">
+                    <a href="/consumenten/disclaimer" class="list__link"
+                        >disclaimer
+                    </a>
+                </li>
+                <li class="list__item">
+                    <a href="/consumenten/privacy" class="list__link"
+                        >privacy
+                    </a>
+                </li>
+                <li class="list__item">
+                    <a href="/consumenten/veiligheid" class="list__link"
+                        >veiligheid
+                    </a>
+                </li>
+                <li class="list__item">
+                    <a href="/consumenten/cookies" class="list__link"
+                        >cookies
+                    </a>
+                </li>
+                <li class="list__item">
+                    <a
+                        href="https://www.wijzeringeldzaken.nl/verzekering/"
+                        class="list__link"
+                        rel="noopener noreferrer"
+                        target="_blank"
+                    >
+                        wijzer in geldzaken
+                    </a>
+                </li>
+            </ul>
+            <div class="main-footer__label-description">
+                <span>Zilveren Kruis is onderdeel van Achmea</span>
+            </div>
+        </div>
+    </footer>
+</template>
